Update - Added new screenshots - Fixed an error when the server was unable to handle requests - Added aggressive mode that can be turned on in configuration - Added a new "db update scanned" command, which updates user names in DB - Modified all service delays in configuration - These features have been added to the configuration "f_rand(min,max)" and "f_rmsec()" for generating random numbers - Turn off SSL detection when connecting to a database - GUI - Added support for unicode and reduced cache for output to window - Enhanced IP scanning. Now it also supports the identification of user names that can be defined in the configuration. And this service requires a connection to the database

+++ m.api/src/main/java/cz/ophite/mimic/vhackos/botnet/api/module/TaskModule.java
@@ -73,7 +73,7 @@ public final class TaskModule extends Module {
     /**
      * Zruší bruteforce.
      */
-    public synchronized TaskResponse abortBruteforce(int bruteId) {
+    public synchronized TaskResponse abortBruteforce(long bruteId) {
         var opcode = new AbortBruteOpcode();
         opcode.setBruteforceId(bruteId);
 
@@ -93,7 +93,7 @@ public final class TaskModule extends Module {
     /**
      * Dokončí bruteforce za netcoins.
      */
-    public synchronized TaskResponse finishBruteforceForNetcoins(int bruteId) {
+    public synchronized TaskResponse finishBruteforceForNetcoins(long bruteId) {
         var opcode = new FinishBruteForNetcoinsOpcode();
         opcode.setBruteforceId(bruteId);
 
@@ -103,7 +103,7 @@ public final class TaskModule extends Module {
     /**
      * Odstraní bruteforce z tásků.
      */
-    public synchronized TaskResponse removeBruteforce(int bruteId) {
+    public synchronized TaskResponse removeBruteforce(long bruteId) {
         var opcode = new RemoveBruteOpcode();
         opcode.setBruteforceId(bruteId);
